+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+Mimic.copy(Ash)
+
 ExUnit.start()